requirements.txt
Flask==2.2.3
# tensorflow==2.13.0
tensorflow-macos==2.13.0
keras==2.13.1
numpy==1.24.3
Pillow==9.5.0
